+++ cargo-web/app/controllers/GameController.scala
@@ -12,8 +12,7 @@ trait NewGameForm {
    val newGameForm = Form(
       tuple (
          "playername" -> nonEmptyText(maxLength = 250),
-         "companyname" -> nonEmptyText(maxLength = 250),
-         "companylocation" -> nonEmptyText(maxLength = 50)
+         "companyname" -> nonEmptyText(maxLength = 250)
       )
    )
 }
@@ -35,7 +34,7 @@ class GameController @Inject() (cc: ControllerComponents)
             logger.warn("New game failed: " + errors)
             BadRequest
          },{
-            case (playerName, companyName, companyLocation) =>
+            case (playerName, companyName) =>
                logger.warn("New game started")
                Ok(views.html.game.welcomePage())
          }
@@ -43,70 +42,22 @@ class GameController @Inject() (cc: ControllerComponents)
    }
 
    def startGame = Action { implicit request =>
-      Redirect(routes.TutorialController.showFirstJob())
+      Redirect(routes.TutorialController.startTutorial())
    }
 
    def continueGame = Action { implicit request =>
       logger.warn("Game continued")
-      Ok(views.html.game.tutorial.firstJob())
-   }
-}
-
-
-@Singleton
-class TutorialController @Inject() (cc: ControllerComponents)
-  extends AbstractController(cc)  with I18nSupport with NewGameForm with WithLogging {
-
-   def showFirstJob = Action { implicit request =>
-      Ok(views.html.game.tutorial.firstJob())
+      Redirect(routes.GameController.showGame())
    }
 
-   def showFirstJobDetails = Action { implicit request =>
-      Ok(views.html.game.tutorial.firstJobDetails())
+   def showGame = Action { implicit request =>
+      Ok(views.html.game.mainWindow())
    }
-
-   def takeFirstJob = Action { implicit request =>
-      Redirect(routes.TutorialController.showFirstTruck())
-   }
-
-   def showFirstTruck = Action { implicit request =>
-      Ok(views.html.game.tutorial.firstTruck())
-   }
-
-   def showFirstTruckDetails = Action { implicit request =>
-      Ok(views.html.game.tutorial.firstTruckDetails())
-   }
-
-   def cantAffordFirstTruck = Action { implicit request =>
-      Ok(views.html.game.tutorial.firstBankLoan())
-   }
-
-   def showFirstBankLoan = Action { implicit request =>
-      Ok(views.html.game.tutorial.firstBankLoan())
-   }
-
-   def showFirstBankLoanDetails = Action { implicit request =>
-      Ok(views.html.game.tutorial.firstBankLoanDetails())
-   }
-
-   def takeOutFirstBankLoan = Action { implicit request =>
-      Redirect(routes.TutorialController.showFirstTruckAgain())
-   }
-
-   def showFirstTruckAgain = Action { implicit request =>
-      Ok(views.html.game.tutorial.firstTruckAgain())
-   }
-
-   def showFirstTruckDetailsAgain = Action { implicit request =>
-      Ok(views.html.game.tutorial.firstTruckDetailsAgain())
-   }
-
-   def buyFirstTruck = TODO
-
-   def garage = TODO
 
    def mechanic = TODO
-
    def driver = TODO
+   def ship = TODO
+   def train = TODO
+   def aeroplane = TODO
 
 }