+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Changelog of changes per release
 
 * 0.01 &alpha;lpha: Not yet released
 * * Added Superhero theme from Bootswatch.com
+* * Static tutorial game flow